+++ controller/ticket.controller.js
@@ -2,8 +2,8 @@ const mysqlQuery = require('../modules/query.module.js');
 const { checkUserticket } = require('../modules/checkUser.module.js');
 const read = (request, response) => {
     const { Cid, Mid, Pid } = request.query;
-    if (typeof (Cid - 1) !== 'number' || typeof (Mid - 1) !== 'number' | typeof (Pid - 1) !== 'number') {
-        response.status(409).send("url query is invalid");
+    if ((isNaN(Cid)) || (isNaN(Mid)) || (isNaN(Pid))) {
+        return response.status(409).send("url query is invalid");
     }
     const sql = "SELECT * FROM `ticket` WHERE `project_idproject` = ? and `customer_idcustomer` = ? and `manager_idmanager` = ?;";
     const data = [Cid, Mid, Pid];
@@ -32,6 +32,9 @@ const create = (request, response) => {
 };
 const update = (request, response) => {
     const id = request.params.id;
+    if (isNaN(id)) {
+        return response.status(409).send("url param is invalid");
+    }
     const { title, description, closedAt, createdAt, status, solution, project_idproject, customer_idcustomer, manager_idmanager } = request.body;
     const sql = "UPDATE `crm`.`ticket` SET `title` = ?, `description` = ?, `createdAt` = ?,`closedAt` = ?, `status` = ?, `solution` = ?, `project_idproject` =?, `customer_idcustomer` = ?, `manager_idmanager` = ? WHERE `idticket` = ?;";
     const data = [title, description, closedAt, createdAt, status, solution, project_idproject, customer_idcustomer, manager_idmanager, id];
@@ -39,6 +42,9 @@ const update = (request, response) => {
 };
 const _delete = (request, response) => {
     const id = request.params.id;
+    if (isNaN(id)) {
+        return response.status(409).send("url param is invalid");
+    }
     const sql = "DELETE FROM ticket WHERE `idticket` = ?";
     const data = [id];
     (checkUserticket(id)) ? response.status(404).send('wrong id') : mysqlQuery(sql, response, 202, "delete", data);
